Séparation des instruction en prévision pour une meilleure lisibilité
