+++ minicoin_inst.h
@@ -58,6 +58,40 @@ typedef struct InstrAffect InstrAffect;
 InstrAffect* newInstrAffect(const char *varname, const Instr *value);
 
 
+#ifdef MINICOIN_INST_C
+#include <stdint.h>
+//#include <inttypes.h>
+
+#define result(typ, res) ((IsrcResult){.typ = res})
+//#define noResult (IsrcResult){.none = 0}
+#define noResult result(none, 0)
+
+void printMarge(const unsigned int sublvl);
+//static inline bool CheckInstrType(const Instr *instr, const InstrType type);
+#define CheckInstrType(instr, ctype) if(instr->type != ctype) {fprintf(stderr, "Erreur %s(%p) n'est pas %s (%d != %d)\n", __func__, instr, #ctype, instr->type, ctype); return false;}
+#define MallocVerif(type, name) type *name = malloc(sizeof(type)); if(name == NULL) { fputs("Erreur malloc(" #type ")", stderr); exit(EXIT_FAILURE); }
+#define internPrint(var, nbsp) (var)->print(var, nbsp)
+#define internCall(var, fn) ((var)->fn(var))
+#define internFree(var) internCall(var, free)
+#define internVerif(var) internCall(var, check)
+#define internEval(var) ((var)->eval(var))
+
+typedef union IsrcResult {
+    char *str;
+    double dbl;
+    uint8_t none;
+} IsrcResult;
+
+struct Instr {
+    InstrType type; //Type d'instruction
+    DataType retour; //Type data retournée par instruction
+    IsrcResult (*eval)(const Instr*);
+    void (*free)(/*const*/ Instr*);
+    void (*print)(const Instr*, const unsigned int);
+    bool (*check)(const Instr*); //true = erreur, false = ok
+};
+#endif // MINICOIN_INST_C
+
 #ifdef __cplusplus
 }
 #endif
